code by https://chat.openai.com/chat
This program includes the cstdlib and ctime libraries, which provide the rand() and srand() functions for generating random numbers and seeding the random number generator, respectively. The time() function is used to seed the random number generator with the current time. This helps ensure that the random numbers generated are truly random, since the seed value will be different each time the program is run.

The rand() function generates a random number between 0 and RAND_MAX (a predefined constant in the cstdlib library). To get a random number between 1 and 10, we use the modulo operator (%) to reduce the range of the random number to the desired range (in this case, 0 to 9). Then, we add 1 to the result to shift the range to 1 to 10.

Finally, we print the random number to the console using std::cout.

I hope this helps! Let me know if you have any questions.